Rework setting tests to use contract trait.
