+++ tests/phpunit/integration/Modules/AdSense/SettingsTest.php
@@ -13,6 +13,7 @@ namespace Google\Site_Kit\Tests\Modules\AdSense;
 use Google\Site_Kit\Context;
 use Google\Site_Kit\Core\Storage\Options;
 use Google\Site_Kit\Modules\AdSense\Settings;
+use Google\Site_Kit\Tests\Core\Storage\Setting_With_Owned_Keys_ContractTests;
 use Google\Site_Kit\Tests\Modules\SettingsTestCase;
 
 /**
@@ -20,6 +21,8 @@ use Google\Site_Kit\Tests\Modules\SettingsTestCase;
  * @group AdSense
  */
 class SettingsTest extends SettingsTestCase {
+
+	use Setting_With_Owned_Keys_ContractTests;
 
 	public function test_register_filters() {
 		$settings = new Settings( new Options( new Context( GOOGLESITEKIT_PLUGIN_MAIN_FILE ) ) );
@@ -107,54 +110,12 @@ class SettingsTest extends SettingsTestCase {
 		$this->assertArrayNotHasKey( 'account_id', $option );
 	}
 
-	public function test_owner_id_is_set() {
-		$user_id = $this->factory()->user->create( array( 'role' => 'administrator' ) );
-		wp_set_current_user( $user_id );
-
-		$settings = new Settings( new Options( new Context( GOOGLESITEKIT_PLUGIN_MAIN_FILE ) ) );
-		$settings->register();
-
-		$fields = array(
-			'accountID' => 'test-account-id',
-			'clientID'  => 'test-client-id',
-		);
-
-		foreach ( $fields as $key => $value ) {
-			delete_option( Settings::OPTION );
-
-			$options = $settings->get();
-			$this->assertTrue( empty( $options['ownerID'] ) );
-
-			$options[ $key ] = $value;
-			$settings->set( $options );
-
-			$options = get_option( Settings::OPTION );
-			$this->assertEquals( $user_id, $options['ownerID'] );
-		}
+	protected function get_testcase() {
+		return $this;
 	}
 
-	public function test_owner_id_is_not_set() {
-		$user_id = $this->factory()->user->create( array( 'role' => 'administrator' ) );
-		wp_set_current_user( $user_id );
-
-		add_option(
-			Settings::OPTION,
-			array(
-				'siteStatus' => 'old-site-status',
-			)
-		);
-
-		$settings = new Settings( new Options( new Context( GOOGLESITEKIT_PLUGIN_MAIN_FILE ) ) );
-		$settings->register();
-
-		$options = $settings->get();
-		$this->assertTrue( empty( $options['ownerID'] ) );
-
-		$options['siteStatus'] = 'new-site-status';
-		$settings->set( $options );
-
-		$options = get_option( Settings::OPTION );
-		$this->assertNotEquals( $user_id, $options['ownerID'] );
+	protected function get_setting_with_owned_keys() {
+		return new Settings( new Options( new Context( GOOGLESITEKIT_PLUGIN_MAIN_FILE ) ) );
 	}
 
 	/**
